some slight refactoring
